Move odcs_util.py into utils/
*OSBS-8605

Also change files which import odcs_utils.py
